+++ x/bank/keeper/migrations.go
@@ -34,3 +34,9 @@ func (m Migrator) Migrate3to4(ctx sdk.Context) error {
 	m.keeper.SetParams(ctx, banktypes.NewParams(oldParams.DefaultSendEnabled))
 	return nil
 }
+
+// Migrate3_V046_4_To_V046_5 fixes migrations from version 2 to for chains based on SDK 0.46.0 - v0.46.4 ONLY.
+// See v046.Migrate_V046_4_To_V046_5 for more details.
+func (m Migrator) Migrate3_V046_4_To_V046_5(ctx sdk.Context) error {
+	return v046.Migrate_V046_4_To_V046_5(ctx.KVStore(m.keeper.storeKey))
+}